Create a separate database class for the database based on Liang Yu's data
